FH-3023 Add asyncGenerateHash function
Currently a large majority of the `JSON.stringify` commands are
made from `generateHash`. In performance analysis of the sync
framework this was found to be an issue.

This adds an `asyncGenerateHash` function and an `asyncGetClientHash`
function to the MBaaS API to allow for these calls to be made
asynchronously.

This also implements these calls internally in the API.