Use new getDitcher act to obtain MongoDB client
This will use the new `getDitcher` act in `fh-db` to set the `mongo`
value in `sync-UpdatesModel_mongo`. At the moment this is doing an
asynchronous call in `setFHDB` and it looks strange.

It may be better to wrap each function in a check for whether
`mongo` has been defined or not, if it's not it will do the
`getDitcher` act and then perform the MongoDB action.

This would ensure `mongo` is set on the first `create`, `list` or
`remove` invocation whereas at the moment there might be a chance
that `mongo` could still be undefined when the first invocation
occurs.