FH-3171 Add hash provider for sync
Adds a module that allows for the hashing algorithm used for each
dataset to be overridden. For each dataset there is a `recordHash`
and a `globalHash` function. The `recordHash` function deals with
the hashing of single records in a dataset. This can be overridden
by calling `setRecordHash` and passing in the ID of the dataset.
The `globalHash` function deals with hashing for an entire dataset,
it accepts an array of record hashes and uses these to calculate
a hash for the entire dataset. This can be overridden in a similar
way to the record hash.