🎨 Try to make logging a little more consistent
JIRA: https://issues.jboss.org/browse/FH-3220

Motivation:

See the conversation in the JIRA issue for more details, but basically
we want to try to achieve some sort of consistency in the logging from
the sync framework.

Modification:

We attempt to follow the suggestions in the bunyan level suggestions
section:

https://github.com/trentm/node-bunyan#level-suggestions

Since this is a library rather than an app, almost all logs are turned
down to "debug" level (bunyan suggest "trace", but not all loggers
have a "trace" level, such as winston, which is what is currently
being used).

Result:

* Most log entries are debug level
* Some log entries are error level, when they add context to an error
  that occurs, and it's difficult to pass that information up to a
  caller.

Additional Notes:

It's not always clear whether we're passing an `Error` object or a
`String` as the *err* first param to a callback. In fact, there's at
least one function here in the sync code which passes an `Error`
object in one case, and a `String` in another case.

That may be something that's difficult to change now. If we could
though, and we choose to pass an `Error` object always, we could
consider using the [verror](https://www.npmjs.com/package/verror)
module to wrap errors with context as they bubble up to the app,
rather than having to log at error level at the points we do.