🐛 FH-3254 Add a null check for dataset client before processing
This change prevents a scenario where a datasetclient has been removed
from the db (via a cleanup task for inactive clients) yet there are
still pending items in the queue for that datasetclient.
In that scenario, it is best to mark the pending update as failed as we
don't have enough info to process it correctly.
The update failed notification will be triggered on the client whenever
they become active again and do a sync with the server