🐛 RH-3754 Fix for sync queue messages
This updates the version fo the fh-sync module to include a fix for the
ttl of queue messages.
See https://github.com/feedhenry/fh-sync/pull/22 for more info.

The dependencies in package.json have also been updated to remove unused
dependencies.
These were discovered when looking at where the mongodb-queue module was
being required. It was originally required in fh-mbaas-api, but was
moved out with the sync implementation to the fh-sync module.
The majority of the unused modules were from fh-sync, with a few extra
ones discovered using this command:

```
dependency-check ./package.json --no-dev --unused
```

The npm-shrinkwrap.json file was regenerated as a result